requirements.txt
django
djangorestframework
django-cors-headers
python-decouple
psycopg2-binary
phonenumbers
djangorestframework-simplejwt
drf-yasg
django-extensions
dj-rest-auth
sphinx
sphinx_rtd_theme
coverage
pytest-django
uritemplate
pyyaml
ipython
django-polymorphic
django-q
redis
croniter
django-bootstrap-datepicker-plus
django-bootstrap4

